MFC r293818
sfxge: cleanup: fix typo in unused EFX_QWORD_IS_SET64

Submitted by:   Richard Houldsworth <rhouldsworth at solarflare.com>
Sponsored by:   Solarflare Communications, Inc.


Former-commit-id: 03b18760b61f2c00386d7b4b4dccb340b93569ab

+++ sys/dev/sfxge/common/efx_types.h
@@ -1179,7 +1179,7 @@ parameter_list|(
 name|_qword
 parameter_list|)
 define|\
-value|(((_qword).eq_u64[0]) == ~((uint32_t)0))
+value|(((_qword).eq_u64[0]) == ~((uint64_t)0))
 define|#
 directive|define
 name|EFX_QWORD_IS_SET32
